My port of CannonML to haxe was done entirely by hand. Given how long that took, I decided to use TiVo's as3tohx to to the initial conversion of this library to haxe.
This did a lot of damage though. Almost all for loops were destroyed
and had to be rebuilt.  I also noticed that some of the AS3 code
depends on the value of the for loop iterator retaining its value
after the loop exits. Haxe doesn't do that and I noticed this problem
very late in my fixes, so there are probably places where this is
still a problem. There are probably places where the wrong iterator is
used or additional iterators were used in the original code but got
dropped.

Also the formatting, especially the comments got totally messed up.
I'm not sure if that tool saved any time or not.

In any case, this checkin is the first that compiles. It doesn't run
yet, but I wanted to get a baseline checkin saved before I started
trying to debug. But the diffs between this checkin and its parent are
a good place to look if you at all suspect something doesn't make
sense or couldn't have worked this way in the original code.